Split test_testing.py (4574 LoC) into separate files
