+++ test/testing/common.py
@@ -1,0 +1,27 @@
+# Copyright 2023 Canonical Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import inspect
+
+
+def get_public_methods(obj):
+    """Get the public attributes of obj to compare to another object."""
+    public = set()
+    members = inspect.getmembers(obj)
+    for name, member in members:
+        if name.startswith('_'):
+            continue
+        if inspect.isfunction(member) or inspect.ismethod(member):
+            public.add(name)
+    return public
